PSM: Use abseil as the source of truth
private-join-and-compute has an older version of Status/StatusOr. It's
being used for all its dependencies. That library does not
have an updated upstream to fetch recent changes too.

PSM libraries (private_membership & SHELL) is depending on it. And now,
PSM libraries migrated the usage of Status/StatusOr to
Abseil only. That makes it much easier to do any sync from the
upstream and have one source of truth for Status/StatusOr type.

That CL is removing the Status/StatusOr helper in the
private-join-and-compute & SHELL libraries. Then, it aliases their usages
to Abseil instead. So we only depend on that
private-join-and-compute on some functionalities that exist,
without depending on its local Status/StatusOr
(like forbidding using ValueOrDie and will be removed in near future).

Note: That migration happened on the upstream for SHELL already. And
it needs to be done locally in Chromium's library version before
retrieving the latest version from upstream. Without these changes in
private-join-and-compute, the latest changes in PSM libraries won't
work. As in the upstream, all of them are using Abseil unlike here.

For more info, see http://go/absl-psm-source-of-truth.

Bug: b:219511651
Change-Id: If40cf20c62bac61a51a23b1c84d542ba59b6f088
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3460585
Reviewed-by: Mirko Bonadei <mbonadei@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Amr Aboelkher <amraboelkher@chromium.org>
Cr-Commit-Position: refs/heads/main@{#971183}
NOKEYCHECK=True
GitOrigin-RevId: 26ada4602cb0a2266554563c31318592f1376a53