[aspect-ratio] Use continued fraction algorithm for LayoutUnit conversion
Previously we'd directly convert the float aspect-ratio to LayoutUnit.
This wasn't correct as we'd potentially lose a lot of precision in the
process.

Instead of this use the continued fraction algorithm to find a precise
representation.

This mitigates the current problem, but likely isn't the best end
scenario. We likely want to represent aspect-ratios as either a
LayoutUnit ratio, or float ratio, depending on the source.

(Specifically the LayoutUnit ratio is far better for the replaced
element case, except for aspect-ratio only SVGs, a float ratio may be
better for aspect-ratio only elements).

Bug: 1296175
Change-Id: I1a1f62fdfdb27785eb076ccaa2d5a4e81566fd46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3459264
Reviewed-by: David Grogan <dgrogan@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#971292}
NOKEYCHECK=True
GitOrigin-RevId: d02f975002dc84d8c69124e9f18943e35be3b54e