[attributon_reporting] Add <img> attributionsrc processing for Register-Source
This CL implements a new surface for Attribution Reporting as documented at:
https://github.com/WICG/conversion-measurement-api/blob/main/EVENT.md#registering-attribution-sources

This API allows an image element to set an attributionsrc=url attribute
which performs a request whose response is allowed to configure
an attribution source.

This only implements the img API for views, which will be extended to
clicks. For this reason, the infrastructure is meant to process and
response to these responses in a keep-alive fashion.

Note: This change adds a fair amount of repetition to processing/parsing
the new AttributionSourceData struct above blink::Impression.

This change reflects the desired end-state, and code which operates on
blink::Impression will be either removed or migrated to the new
pending_receiver<> design.

Bug: 1275201
Change-Id: Ib8787e65f3ed1538081e90caa45edb79fe738d14
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3433532
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Nan Lin <linnan@chromium.org>
Reviewed-by: Andrew Paseltiner <apaseltiner@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Commit-Queue: John Delaney <johnidel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#971338}
NOKEYCHECK=True
GitOrigin-RevId: 2fae05457e9b1bd833a51e46b9d37cf3dd0fe0d4