Add createImageBitmap and drawImage tests for ImageDecoder frames.
This adds createImageBitmap tests for ImageDecoder, VideoDecoder,
and canvas backed frames as well as a drawImage test for an
ImageDecoder backed frame.

This ensures we have coverage of SkImage unwrapping paths as well
as general coverage for createImageBitmap.

Even though createImageBitmap(videoFrameFromCanvas) is unlikely to
be common, it ensure coverage of texture backed VideoFrame
to ImageBitmap.

R=sandersd

Bug: 1275360
Change-Id: I415d4acf9a9927307aa0e5b776a6a8541caa652d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3466581
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#971565}
NOKEYCHECK=True
GitOrigin-RevId: 3f176bacec454ec9e78ffdb5ec2e122dc4675558