Fix Canvas Null-dereference READ
The canvas context created from the iframe has a font that is created
with the iframe's FontSelector. When the iframe is removed
and garbage collected along with its FontSelector, the
|font_selector_map_| in the font fallback list, which is a weak
persistent, is garbage collected and set to null. In our code, we forget
to check if FontSelector is valid before using it.

Bug: 1290091, 1289852

Change-Id: I82ef3e967037d8e5a40a056610ee6cf3fbf15045
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3461698
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Commit-Queue: Yi Xu <yiyix@chromium.org>
Cr-Commit-Position: refs/heads/main@{#971674}
NOKEYCHECK=True
GitOrigin-RevId: 272b5519bcc3b4023b03ce3fdef26f8ea6832f1b