Update body overflow on DOM mutations
Inserting or removing a body element may change which is the first body
element. That is, the body element which contributes as the viewport
defining element for propagating scrollbars, backgrounds, etc to the
viewport. StyleEngine::ViewportDefiningElementDidChange() already
handled the case where the viewport defining element change between body
and the html root element, but did not handle the case with insertion or
removal in the case of multiple body elements.

Bug: 1296166
Change-Id: I03d41d8d94a4212e5f0e56712519875caf9ecffd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3466114
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#971703}
NOKEYCHECK=True
GitOrigin-RevId: 41af630242bc38a7c27399699a88396619335448