Fix flaky MetronomeLikeTaskQueueTest.PostTaskRunsPriorToTick test.
As I was running gtest-parallel, I noticed that this test was failing
~5 times out of 100 runs due to the callback having been called prior
to task_environment_.FastForwardBy(). I discovered that this is due to
the test task environment being multi-threaded and non-delayed PostTask
tasks are able to run without advancing time. We still want to advance
the time though to ensure that the task always have had time to run.

The fix to the flake is to move the "EXPECT_FALSE" call up by one line.

Bug: chromium:1253787
Change-Id: Ifd5fab71e42467f7686018d35a7981b6a7c87692
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3466056
Reviewed-by: Mirko Bonadei <mbonadei@chromium.org>
Commit-Queue: Henrik Boström <hbos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#971715}
NOKEYCHECK=True
GitOrigin-RevId: 8eeb68ce706c0f487b273daf7c15f6b14d663560