Create text fragment handler and generator on demand
Changes in this CL:
- As TextFragmentHandler can be created not only for generation but also
for reshare and remove share, don't create generator in the constructor
of TextFragmentHandler but only when text fragment needs to be generated
upon context menu open.
- Create TextFragmentHandler only when needed instead of always creating
it on main frame init.
- After generation result is served by callback to the browser, the
generator can be safely reset. There is no good indicator when generator
can be reset in cases when browser never requests a selector.

Bug: 1290430
Change-Id: Ia5c97c057c059fe75522f50a664cc206a5c7f812
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3462230
Reviewed-by: David Bokan <bokan@chromium.org>
Reviewed-by: Cheick Cisse <cheickcisse@google.com>
Commit-Queue: Gayane Petrosyan <gayane@google.com>
Cr-Commit-Position: refs/heads/main@{#971812}
NOKEYCHECK=True
GitOrigin-RevId: efb3ee9e39cefeffb228c6efa5768a9b129bd0c9