[FlexNG] Support forced break-before/after for flex rows
This CL adds basic support for break-before/after for flex rows.
Similar to CL:3287251, we store the BreakBetween value of each row
on the break token (with the last BreakBetween value representing
the final break-after to be propagated to the container).

BreakBeforeRowIfNeeded() was added to determine whether or not to
break before a given row. If we decide to break before a row, we will
add a break before the first item in that row (since rows themselves
don't produce a layout result).

More complex scenarios will be supported in follow-ups (as well as
break-before/after:avoid).

Propagation of the initial row break-before and final row break-after
to the container will also be done in a follow-up CL.

Bug: 660611
Change-Id: Ib5f04916d8119cb28dd4d5388af7a3101e894285
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3466756
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#971960}
NOKEYCHECK=True
GitOrigin-RevId: c438e21a5f71eb87df99805e07a1b559533f176d