serial: Use AbortSignal to abort write()
This change adds an algorithm to the AbortSignal from the
WritableStreamDefaultController which causes the pending operation
created by a call to write() to be aborted. This avoids a situation
where a queued write operation will prevent the port from being closed
until it has completed.

Bug: 1295963
Change-Id: I5e894eb51e135cfb663276f579b57f117295242c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3366788
Reviewed-by: Adam Rice <ricea@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#971977}
NOKEYCHECK=True
GitOrigin-RevId: 848f35dab97b297664a4e42b18ddbd62ededf909