[GridFragmentation] Propagate break-before/break-after from container.
...when unfragmented). This patch captures the break-before/break-after
values for grid row(s) when unfragmented. We pass in a break-between
vector into |PlaceItems| which will then use this to set the
break-before/break-after on the fragment.

We also use these values to add the correct break-before/break-after
on the grid container.

Bug: 614667
Change-Id: Ie529efe0e5bfe53262d10d874149ffc5c6e67100
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3453286
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972206}
NOKEYCHECK=True
GitOrigin-RevId: 5c52e95680d27cab396f79b9db9461f0e8362cfa