Change GPUDeviceLostInfo.reason to any
This CL changes the type of GPUDeviceLostInfo.label in IDL to follow the
latest WebGPU spec. Ideally the type should be
(GPUDeviceLostReason or undefined) for the consistency with the spec but
currently the bindings somehow can't handle it although it should be
vaild. Until the bindings supports it we set the type as any and do
custom handling in Blink code as a workaround.

Bug: 1290729
Change-Id: I298738fd502c2aab6bcaf6cbb2cc66a023181e60
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3455182
Reviewed-by: Brandon Jones <bajones@chromium.org>
Reviewed-by: Austin Eng <enga@chromium.org>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972240}
NOKEYCHECK=True
GitOrigin-RevId: 64eef749cabe22b649f27cbebc02a3c594e54c10