[Metronome] Fix issue where low precision tasks run out of order.
After refactoring WebRtcMetronomeTaskQueue to use PostDelayedTaskAt()
we accidentally broke the behavior that delayed tasks that happen on the
same tick should still be executed in the order of their target
execution time. The snapped times may be identical, but the original
target times are not and should be respected.

The test "DelayedTasksRunInOrder" did not catch this on the CQ due to
one of the delays in the test being zero and triggering a non-delayed
path under the hood, but this caused flaky behaviors on the bots. The
test is updated with non-zero delays which made reproducing the issue
consistent.

This CL fixes the ordering issue of both WebRtcMetronomeTaskQueue and
ThreadWrapper by introducing a helper class that keeps track of
coalesced tasks until they are ready to run: CoalescedTasks.

The TQ implementations are still responsible for scheduling using
PostDelayedTaskAt(), CoalescedTasks is only responsible for ordering
and keeping track of whether or not scheduling is needed for newly
enqueued tasks.

Bug: chromium:1297095
Change-Id: Ica597154fa85976f61a1c69066b52e8ab120c9c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3465278
Reviewed-by: Mirko Bonadei <mbonadei@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Evan Shrubsole <eshr@google.com>
Commit-Queue: Henrik Boström <hbos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972417}
NOKEYCHECK=True
GitOrigin-RevId: a9582dc504511cca361d708489e99996b9ad6b87