Make sure we have consistent ComputedStyle flat trees
An element in the flat tree should never have a ComputedStyle if the
flat tree parent does not have one. When doing slot re-assignments we
could end up having ComputedStyle on a host child which was moved from a
slot which was part of the rendered flat tree into a slot that was
either in a display:none subtree or even outside the flat tree because
of another host ancestor without a slot in its shadow tree.

Make sure we clear the ComputedStyle, detach the layout tree, and clear
dirty bits in such cases.

Bug: 1282460
Change-Id: I8c3280f7ee580d6197ce8f5507150177639143d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3468706
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972454}
NOKEYCHECK=True
GitOrigin-RevId: 52445a4616e86c41eff2ca001ea844befdb5ff39