SET: Remove the viz implementation from the renderer
This patch removes the viz side of the implementation from the
renderer code. The cc/viz code is still there, to be cleaned up
in separate patches. The reason for the clean-up here now is that the
code will evolve soon and it's easier to only maintain one branch.

R=pdr@chromium.org

Change-Id: Ib3848d92d59b48ded8c49cd85bc2a070b8038349
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3469229
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Vladimir Levin <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972689}
NOKEYCHECK=True
GitOrigin-RevId: f9664aa2432c9c8b3501782d7500c3aa9938e640