AppHistory bfcache support
Currently, with appHistory and bfcache both enabled, appHistory's
entries() arrays are not updated when a navigation is served from
bfcache. It keeps the same appHistory.entries() that it had when the
document was put into bfcache, but the entries() almost certainly have
changed in the meantime.

This CL ensures that when blink is notified of the restore from
bfcache, it also receives updated appHistory.entries(). Any entries
that evicted from appHistory.entries() as a result of the restore
will have dispose events fired.

Fixed: 1241655
Change-Id: I9c49312ccbbec714f751f3b59c8d1f05edb024fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3277336
Reviewed-by: Charles Reis <creis@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Domenic Denicola <domenic@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Nate Chapin <japhet@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972761}
NOKEYCHECK=True
GitOrigin-RevId: 97d2f54ec74376973d170715453c477af3badc42