Implement JS Origin Trial for partitioned cookies
This CL implements the Origin Trial for partitioned cookies (aka CHIPS)
accessed via JavaScript APIs.

The CL adds new boolean arguments to RestrictedCookieManagers which
renderers can pass to the browser which filters partitioned cookies
if the caller is not enrolled in the OT.

If a renderer were to be compromised and opt their frame into the trial
when they shouldn't be, they would be accessing cookies which are more
restricted than SameSite=None cookies today.

Bug: 1225444,1296161
Change-Id: I6f177c1528161c36294dffb4328801e62abb099a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3470554
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Bo Liu <boliu@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Commit-Queue: Dylan Cutler <dylancutler@google.com>
Cr-Commit-Position: refs/heads/main@{#972767}
NOKEYCHECK=True
GitOrigin-RevId: 49dc3925fce437ffdd71912bc7091d274bd162fd