Make CSSToStyleMap::MapNinePieceImageRepeat accept non-pair values.
Properties such as border-image-repeat are ostensibly pairs
of vertical/horizontal, but as a shorthand, the user can specify
only one of the values. CSSToStyleMap::MapNinePieceImageRepeat() silently
discards the request if the value given is not a pair, which is bad if
we're trying to give in border-image-repeat's initial value (which is
a CSSIdentifierValue, not a CSSValuePair, and thus needs expansion).

Change-Id: If939136a4fa365e75e9a27dc8ad859ad67471cd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3468767
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972931}
NOKEYCHECK=True
GitOrigin-RevId: 3513c3deb50aef33d4f3261b38559a37d3f68d46