[inert] Re-enable a11y dump tree test for inert attribute
I suspect the test was flaky because LocalFrame::SetIsInert was not
clearing the accessibility tree cache when changing the inertness.
Doing so aligns it with FullscreenElementChanged() from fullscreen.cc
and with InertSubtreesChanged() from html_dialog_element.cc.

Also, the test wasn't really testing the effects of inertness on
focusability due to a typo in the JS that was supposed to make all
elements potentially focusable.

Bug: 1298003, 692360
Change-Id: I3ee7cf3668950034573faf0844f2bc60d56b6732
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3468495
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Cr-Commit-Position: refs/heads/main@{#972964}
NOKEYCHECK=True
GitOrigin-RevId: 24a066cc5047558adbd85fdc562b3b46d23367db