[cropToPromise][#4] BadMessage on non-increasing crop_version
Issue a BadMessage if the render process provides a non-increasing
crop-version to the browser process.

Because WebContentsFrameTracker is the first entity to know that
the crop-version is non-increasing, it starts the process of issuing
the BadMessage. However, as WebContentsFrameTracker does not itself
know the render process of the capturing render process, which is
the compromised process, WCFT informs MediaStreamDispatcherHost of
the issue, and MSDH issues the BadMessage on WCFT's behalf.

Note: Currently, cropping is only supported for self-capture,
so one could have thought that we can just kill the capturer/capturee,
as they are one and the same. However, this would have been a mistake,
as the document that holds the track and initiates the crop,
is not necessarily the top-level document.

Bug: 1266378, 1299110
Change-Id: Ib94ea2e83cce97b63a03c7c5fd59b8367475a6e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3472068
Reviewed-by: Jordan Bayles <jophba@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Bo Liu <boliu@chromium.org>
Commit-Queue: Elad Alon <eladalon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#973223}
NOKEYCHECK=True
GitOrigin-RevId: 96689d9620cb0eae8f49f65f4aa15c095b19100d