[FlexNG] Column balancing with forced breaks
Because we didn't handle expansions when breaking before during the
initial column balancing pass, we could end up with items or rows
having a negative offset resulting in an inaccurate column block
size.

Move the expansion handling from where the break before is added to
where the break before is handled in the next column. We know the
consumed block size in the subsequent column, and the consumed block
size can be used to determine the amount to expand by, even if the
final fragmentainer block size is unknown.

Bug: 660611
Change-Id: I1309b3906126fc9a838634596cd6280241d5efee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3475053
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#973795}
NOKEYCHECK=True
GitOrigin-RevId: 7b45bc104d43553b0c768ae8bbc8b8ed4f4edf49