Miss the cache on nested multicol with OOF.
We handle OOFs that are contained by a multicol in a nested
fragmentation context in a special way: We'll lay them out when we have
returned to the outermost fragmentation context root (that's when we
know how many inner fragmentainers we've got). If we both do this, and
hit the cache, and thus keep and re-use the old OOF fragments along with
the rest, we'll end up with extraneous fragments when we return to the
outermost fragmentation context root and add fragments for them there as
well.

Bug: 1296900
Change-Id: I78193353092529e3bd11d58785afa8e4d1f19ebb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3481336
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#973852}
NOKEYCHECK=True
GitOrigin-RevId: 3079350de239212574debb3aeb9d3b2c1207951e