An unassignable slot should just be treated like a normal DOM node
A <slot> that is not in a shadow root should be treated like any other
element. Almost any code that just checks for a slot element without
using ToHTMLSlotElementIfSupportsAssignmentOrNull()is probably
incorrect.

This seems to help improve performance in issue 1281936, but it's not
clear why. However, it is more correct code and we should land it
whether it improves performance at all. We should not be unnecessarily
truncating subtrees that use a <slot> element

Bug: 1281936
Change-Id: Icb954d8706c3dafc0aaa1fca8d4f3fb9d8c2b44d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3472410
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#973886}
NOKEYCHECK=True
GitOrigin-RevId: 3ff4325989fdbac29152e3e1a57d496416ac2559