Re-write the picker-common.js helper plus *all* of the associated tests
****** SHERIFFS PLEASE READ! ************************************
Help me help you! This is a very large CL that took many patchsets
to get under control. If possible, it would be great if you avoided
reverting this CL. If a few of the tests touched by this CL turn out
to be flaky, please just disable those tests and assign any bugs to
me, and I will take a look. The entire point of this CL is to reduce
test flakiness, so hopefully it can stay landed. :-)
*****************************************************************

All of these picker tests are flaky. They also use quite old test
technology. This CL re-writes the infrastructure for testing the
appearance of pop up pickers, plus all of the associated tests.
Hopefully all of this work was worth it and the flakiness goes away.

This is a large CL. The important changes are in picker-common.js,
which is completely re-written. Most of the test files that use
picker-common were also re-written, primarily to simplify them.
There is a short list of problematic tests in TestExpectations, which
I need to work on after this CL lands, and that work is tracked in
crbug.com/1299212.

See the comments I've placed on this CL - I very carefully went through
all of the new/changed baselines, and made sure they all make sense.
It would be great to spot check these if possible, but there are many
files (>1000) so perhaps focus on the changes to the tests (.HTML)
rather than the baselines (.PNG or .TXT).

Fixed: 1285304,1293965,1293878,1295712,1295733,1264802,1270510
Bug: 1299212
Change-Id: I15009bf95dc52549268df37932637d8ece516999
Cq-Do-Not-Cancel-Tryjobs: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3454445
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#973914}
NOKEYCHECK=True
GitOrigin-RevId: cfec2f83a152029913a4a08eb539a1759dd1cecd