Avoid excessive work for potential name changes in large subtrees
While investigating a large performance degradation in Salesforce,
it was discovered that tons of extra serializations were being
cause by code that recomputes the name on containers. The page
in question had tabindex=-1 on several large containers, including
the <body>. This in turn caused the body to be marked dirty many
times. Each time the serialization loop ran into the body, it
called SerializeChanges(), and ended up with a large list of
serialized nodes which now needed to be de-duped against the
entire set of changed nodes.

This CL avoids reserializing the body, as well as puts other
reasonable limits in the name refreshing algorithm.
Note that other complementary fixes are being worked on as well.

Bug: 1281936
Change-Id: Ib1592be356f5c2c45001b7a940c8dd5fa38314a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3472748
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#973939}
NOKEYCHECK=True
GitOrigin-RevId: 514c466d91241295d0c13f6ebc86cd814e9315db