Add internal pseudo class for SelectorFragmentAnchor
We are using :target pseudo-class for styling the element in SelectorFragmentAnchor, so developers can customize the styling as well.
SelectorFragmentAnchor explainer: https://github.com/WICG/scroll-to-text-fragment/blob/main/EXTENSIONS.md#proposed-solution

When the SelectorFragmentAnchor becomes active, we set CSS target to be the first element that matches the CSS selector that is specified in the URL.
https://example.org#:~:selector(type=CssSelector,value=img[src$="example.org"])
In the above example, the CSS target will be set to the element that matches img[src$="example.org"].

But using only :target is not enough, because we want the style to change only in CssSelectorFragment mode, not ElementFragment, not TextFragment (we are using :tagret in all of them).
So we have introduced :-internal-selector-fragment-anchor, and added some default styling for it to html.css.

A method named SetSelectorFragmentAnchorCSSTarget is added to document.h. Since it is supposed to be called only from CssSelectorFragmentAnchor, it calls PseudoStateChanged for both kPseudoTarget and kPseudoSelectorFragmentAnchor.

Bug: 1252460
Change-Id: Ie7469eecbfc38fc17342512f93ac14170c9560ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3404932
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Mehdi Kazemi <mehdika@chromium.org>
Cr-Commit-Position: refs/heads/main@{#973978}
NOKEYCHECK=True
GitOrigin-RevId: e3f84fa00b53754c78df3f7db06291500757111d