Add order to cascade layers
This change adds a field `order_` to cascade layers. The field is used
in the cascade layer map to assign a cascade priority for a given tree
scope. Assigning the ordering directly on the layer makes it easier for
DevTools to extract ordering information about layers.

Bug: 1240596
Change-Id: I6cd987486d2f2f1344ec805ef417eb6a2894b20e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3482315
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Commit-Queue: Johan Bay <jobay@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974075}
NOKEYCHECK=True
GitOrigin-RevId: 3c0e7c4ccab7523d580b453520608de09b32e7c9