Support sibling combinators and complex selector in :has()
To invalidate the subject elements of a style rule which contains
sibling combinator or complex selector in :has() pseudo class, added
new direction of walking up previous siblings in SelectorChecker
and StyleEngine.

To prevent unnecessary previous sibling traversal or invalidation,
'siblings_affected_by_has_' flag was added to the ElementRareData
to indicate that a previous sibling of the element may match a :has()
pseudo class.

And the flag 'ancestors_affected_by_has_' was renamed to
'ancestors_or_ancestor_siblings_affected_by_has_' to have a clear
meaning. It indicate that a ancestor or a ancestor's sibling of the
element may match a :has() pseudo class.

Bug: 669058, 1276110, 1282294, 1282826, 1284204, 1287680
Change-Id: I4f11c411a7c66b5892f634db60f5bba16470b11b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3455366
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Byungwoo Lee <blee@igalia.com>
Cr-Commit-Position: refs/heads/main@{#974116}
NOKEYCHECK=True
GitOrigin-RevId: aa134e60af012ea65e88e05ec73ca68a5ae37ca1