Migrate Client Hint APIs to Origins
The spec is defined exclusively in origins and its generally not
recommended to use URLs in place of origins. The main culprit is
ClientHintsControllerDelegate::GetAllowedClientHintsFromSource, and the
rest of the changes are essentially refactoring the rest of client hints
code to match.

In areas where the navigation has not been committed yet, the origin
will be taken from the current URL. Otherwise, the origin will be taken
as it normally would (e.g. through GetLastCommittedOrigin())

Bug: 1289324
Change-Id: I9fb570882abf50e52404885b96271368de1eb7ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3472751
Commit-Queue: Aaron Tagliaboschi <aarontag@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Ali Beyad <abeyad@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974234}
NOKEYCHECK=True
GitOrigin-RevId: 579c632e672b4cb4cd7768daa7da5b2f897215f5