webauthn: drop kWebAuth feature flag.
This was added when WebAuthn was under development. It has been
default-enabled for years now but still serves two purposes:

  1) It is disabled it to disable WebAuthn in WebViews.
  2) It is disabled on Android prior to N, because WebAuthn isn't
     supported by Play Services in that case.

We want to add optional WebAuthn support for WebViews. That requires a
more flexible control such as Blink preferences. Thus this change
removes the `kWebAuth` feature and replaces it with a preference that
can disable WebAuthn.

It's worth noting that disabling WebAuthn was never a security barrier,
it's just so that websites doing feature detection get the correct
result. It doesn't matter that a compromised renderer could send
WebAuthn requests to the browser even if it was “disabled”.

BUG=1284805

Change-Id: If32212d4eb620a9336386f98d87f0096e79c6bba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3466764
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974302}
NOKEYCHECK=True
GitOrigin-RevId: 6d11e3a3a1eb22f5c2ddaea8d9fc7fc8eb07acc9