InterpolableLength::CreateLength should create a fixed length when without percentage
When an InterpolableLength is an expression without percentage, it can
still create a fixed length when a CSSToLengthConversionData is
provided.

This patch fixes the bug that an unresolved length (represented as an
expression with pixels, percentages and operations) is returned in this
situation.

Fixed: 1299361
Change-Id: Ib9ca8c465a495fdd0d1d00f5b0d8b6f3b62c5d02
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3485581
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974384}
NOKEYCHECK=True
GitOrigin-RevId: d019371e778ac0c708a91f8d40b03bbe006d2163