BFCache: Support BFCache for Dedicated Workers on non-PlzDedicatedWorker
This CL supports BFCache for Dedicated Workers on
non-PlzDedicatedWorker. For PlzDedicatedWorker, it's already supported.

> Background

In the current implementation, BFCache for Dedicated Workers depends
on PlzDedicatedWorker. At first we thought that we are able to launch
the PlzDedicatedWorker soon, but it turned out that we need more work
for that. To avoid the BFCache from being blocked by PlzDedicatedWorker,
this CL makes it possible to run BFCache on non-PlzDedicatedWorker.

> Design

The infra of BFCache for Dedicated Workers has already been implemented.
To use it for non-PlzDedicatedWorker, mojo connection between the
renderer-side worker context and the browser-side BFCache backend has to
be established.

To achieve it, this CL passes mojo::PendingRemote of
mojom::BackForwardCacheControllerHost from content::DedicatedWorkerHost
in the browser process to blink::DedicatedWorker in a renderer process
during worker startup.

Change-Id: I6c94d08913ead2fb2bd79e82a3bf07f131f927b4
Bug: 1146955
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3469119
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Asami Doi <asamidoi@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974554}
NOKEYCHECK=True
GitOrigin-RevId: 379de7e5e8249a60ed1c4f6e14558e378aee2f35