[tables] Disable the table border quirk.
This disables the logic for the table border quirk. See:
https://github.com/w3c/csswg-drafts/issues/6230
For more information.

This doesn't remove the logic - however the primary function which uses
this:
NGTableBorders::GetCollapsedBorderVisualSizeDiff
will return an empty NGBoxStrut now.

If this lands successfully we'll remove the remaining logic in:
https://chromium-review.googlesource.com/c/chromium/src/+/3484918/1

Bug: 1238243
Change-Id: Ic5c8a369b09cc693e4a4b20f269ccd5a159c3e98
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3485478
Reviewed-by: Aleks Totic <atotic@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974738}
NOKEYCHECK=True
GitOrigin-RevId: 0da4bca813208acafeb242b108b1f1791a960c8f