[FlexNG] AttemptSoftBreak() for rows
Add an optional block-size parameter to AttemptSoftBreak() and
BreakBeforeChild() so that the flex row block-size is used when
breaking before a row. As a result, AttemptRowSoftBreak() can
be removed.

As a result of this change, BreakBeforeChild() will now use the
correct block-size for rows when calling PropagateSpaceShortage(),
which results in the following test passing as expected:
external/wpt/css/css-break/flexbox/multi-line-row-flex-fragmentation-038.html

Bug: 660611
Change-Id: Icaa3ddb4b92298da3ae3e641fe5cc45565048543
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3486261
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#974855}
NOKEYCHECK=True
GitOrigin-RevId: a04a8813adba6729f311d4db29b794ecf837b377