Ensure correct callback order in media/video-poster-after-playing.html
It's theoretically possible that the `requestVideoFrameCallback`
callback could be scheduled to run before the `onplaying` callback,
defeating the purpose of the test. This CL updates the test so that the
ordering is respected.

This is a followup to the final comments in
https://chromium-review.googlesource.com/c/chromium/src/+/3486098

Bug: 1281780
Change-Id: Ie435f26976e0a66c8905ed123cc1a206b85207d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3489916
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Will Cassella <cassew@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974955}
NOKEYCHECK=True
GitOrigin-RevId: 08ad2cadd41f22479de20d0c66b1d7944916e605