Avoid ComputeBlockSizeForFragment() on table sections / rows.
We used to do this in NGSimplifiedLayoutAlgorithm (but ignored the
result). ComputeBlockSizeForFragment() isn't called by the respective
row / section layout algorithms, so let's refrain from doing that in the
simplified layout algorithm as well. The problem was that no intrinsic
block-size is stored in the layout result for rows and sections, and
we'd fail a DCHECK in ResolveBlockLengthInternal(), since we were
passing an intrinsic block-size (0) less than border+padding.

Note that this CL will also skip ComputeBlockSizeForFragment() for table
containers. It should be harmless to keep that, but since we're ignoring
the resulting block-size for tables as well, it seemed reasonable
(rather than making the debug code look silly).

This is a speculative fix for crbug.com/1289408

Bug: 1289408
Change-Id: Ic817c67342891cd774e335c7560be2dfe915fb8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3487914
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#975062}
NOKEYCHECK=True
GitOrigin-RevId: 3cd58e8035aced873e400cd8c1ebe76e7cf97994