Avoid CSSLengthArray in LengthUnitsChecker
It is not necessary to keep an entire CSSLengthArray in
LengthUnitsChecker, since we only ever need to do a linear traversal
of the actually present units (and not random access).

This CL uses a Vector of unit/length pairs instead.

This prepares for a subsequent change that should address a
performance regression caused by increasing kLengthUnitTypeCount
(Issue 1294626).

Bug: 1093055, 1294626
Change-Id: Icd5372680fac2323d38568679c37fc8e6483f121
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3480815
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#975073}
NOKEYCHECK=True
GitOrigin-RevId: 91f2ac1f9d775085fc2b66ad7f5530ba5e6b540b