Refactor SameSiteCookieIssue to CookieIssue
These changes will allow issues to be shown in devtools when a cookie
has one or more warning or exclusion reasons associated with it. It
refactors the code path used for same site cookie issues by making it
available for any type of cookie issue.

Additional changes will need to be made in devtools frontend to account
for the changes introduced here. This will be done in a followup CL.

Bug: 1254192
Change-Id: I929435567eac90bed60c4df464150db6c456da83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3449712
Reviewed-by: Steven Bingler <bingler@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Juba Borgohain <chromiumjuba@gmail.com>
Cr-Commit-Position: refs/heads/main@{#975286}
NOKEYCHECK=True
GitOrigin-RevId: c93969e60c550dcd356e12a91d8530d06926c641