Refactor AtomicStringTable WeakFind* APIs.
The AtomicStringTable::WeakFind* APIs where introduced to allow for find
operations on the AtomicStringTable without actually taking references
on the underlying StringImpl. However, many of these methods are not
being used or are too general and perform redundant checks.

This CL performs the following refactor:
  1. Remove unused methods.
  2. Make Element::WeakLowercaseIfNeeded receive |AtomicStringImpl&|
     thus eliminating unnecessary `isAtomic()` checks.
  3. Precompute the hashes that will be used for the find operations
     in the table, so we can keep them out of the critical section
     when we introduce locks into the AtomicStringTable.

Latest pinpoint show an improvement of 12% in get-attribute and 11.5%
in get-attribute-rare microbenchmarks, which are known users of
this path.

Bug: 1083392
Change-Id: I1e1d3cab24e0089dc5fad785c1b95a30d121fd08
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3491115
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Luis Fernando Pardo Sixtos <lpardosixtos@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#975663}
NOKEYCHECK=True
GitOrigin-RevId: 38ed677e80f9ee253ec77aab67e44a4251f4cbd9