Fix no-hover-during-smooth-keyboard-scroll test
This patch is step 1 in updating the no-hover-on-___-scroll tests.
Previously, the test did not actually validate that the hover state
remained consistent during the scroll. Furthermore, it used a sequence
of scrolls each separated by a pause to wait for the scroll to complete.
This test was occasionally timing out, presumably because the waits were
consuming an appreciable part of the time budget for the test.

This patch adds a scroll listener that verifies the hover state is not
changing, and uses page down instead of arrow down so that a single
scroll can be used.

If this patch address the flake, then a follow up patch can update the
remaining no-hover tests since they have the same issue with not
actually testing that the hover state remains consistent throughout the
scroll.

Bug: 1229064
Change-Id: I14cb491c402ff692a740c7e8229526b24965965a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3488537
Reviewed-by: Steve Kobes <skobes@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#975814}
NOKEYCHECK=True
GitOrigin-RevId: dbf702392b30c01378cee671ea271592e1708bf3