[FlexNG] Flex row breaks when there is no container separation
Handle IsBreakableAtStartOfResumedContainer() for items in a row.
As a part of this CL, we now pass in a bool to
BreakBeforeChildIfNeeded() to tell whether the child is an item in a
row so that we only perform the code needed for such items. For
example, break-before/break-after values should have no effect on
row items (since that is taken care of by the row itself).

Note: multi-line-row-flex-fragmentation-041.tentative.html and
multi-line-row-flex-fragmentation-042.tentative.html worked previously
to this CL but were added to ensure those cases were working
without any additional handling at the row level.

Bug: 660611
Change-Id: I755428b5bf3749c21819ebfaca53efdb725f467a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3491562
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#975965}
NOKEYCHECK=True
GitOrigin-RevId: 41316e5b20a417a391cdab0e8118ff08219293f2