[attribution_reporting] Process event trigger header for attributionsrc
Adds processing for the event trigger header as documented in:
https://github.com/WICG/conversion-measurement-api/blob/main/EVENT.md#triggering-attribution

This change only parses and plumbs the header to the browser process.
Browser process registration of the source with the manager is left as a
followup.

Bug: 1294286
Change-Id: Ibaf8bd7a1b61b3c97c8a912e89919a0fa7cf6ed0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3473072
Reviewed-by: Nan Lin <linnan@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Andrew Paseltiner <apaseltiner@chromium.org>
Auto-Submit: John Delaney <johnidel@chromium.org>
Commit-Queue: John Delaney <johnidel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#976558}
NOKEYCHECK=True
GitOrigin-RevId: faf5b2165c8c2dc1c8a268888a5705ca47b3ee97